Move search views into separate file
